// FutureCounselor/src/app/Models/signup.ts
export class signup {
    email: string;
    password: string;
    rePassword: string;
    firstname: string;
    lastname: string;
    location: number;
    qualification: number;
    percentage: number;
    annualBudget: number;
  
    constructor() {
    }
};
export class option {
    id: number;
    vlaue: string;

    constructor() {
    }
};